django_backend/requirements.txt
# Django Backend Requirements for Hospital Management System

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
sendgrid==6.10.0
twilio==8.10.0
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0

# Development dependencies
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
coverage==7.3.2
black==23.10.1
flake8==6.1.0